+++ packages/terra-form-select/src/_Frame.jsx
@@ -11,6 +11,10 @@ const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
+   * Text for the clear option.
+   */
+  clearOptionDisplay: PropTypes.string,
+  /**
    * Whether the select is disabled.
    */
   disabled: PropTypes.bool,
@@ -89,6 +93,7 @@ const propTypes = {
 };
 
 const defaultProps = {
+  clearOptionDisplay: undefined,
   disabled: false,
   dropdown: undefined,
   dropdownAttrs: undefined,
@@ -363,6 +368,7 @@ class Frame extends React.Component {
 
   render() {
     const {
+      clearOptionDisplay,
       disabled,
       display,
       dropdown,
@@ -441,6 +447,7 @@ class Frame extends React.Component {
                  onRequestClose: this.closeDropdown,
                  searchValue: this.state.searchValue,
                  maxSelectionCount,
+                 clearOptionDisplay,
                })}
           </Dropdown>
           )