+++ packages/terra-form-select/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Added
 * Added option to limit the maximum number of selection count for multi-select variants.
+* Introduce an option to allow clearing the currently selected item
 
 5.8.0 - (March 5th, 2019)
 ------------------